Added Jerry James patches
git-svn-id: svn+ssh://spartan.csl.sri.com/svn/public/pvs/trunk@5329 ab98feae-9515-0410-836f-c1e8bd48b6aa

+++ BDD/bdd/src/bdd_vfns.c
@@ -40,7 +40,7 @@ BDDPTR *MakeBDDVec (int size)
     BDDPTR *F;
 
     F = CALLOC_ARRAY (size+1, BDDPTR);
-    F[0] = (BDDPTR) size;
+    F[0] = (BDDPTR) (long) size;
     return F+1;
   }
   return NULL;
